RTF: Added character escapes and (hackish) encoding support.
RTF does not really support character encodings, but it does
support a form of unicode. If an encoding is declared, t2t
will attempt to convert this into unicode before escaping
the text for Rtf readers to use.

In Rtf readers that do not support Unicode, all unicode characters
will appear as '?'

Not all encodings are supported. In particular, some mutli-byte
encodings (Shift-JIS is one known offender) will use byte 0x0D
as part of a character, but t2t will interpret this as a newline.
This will usually cause an exception to be thrown and the script
to abort.

If there is no encoding declared, t2t should assume the text is
plain ascii and pass it through untouched.



git-svn-id: http://txt2tags.googlecode.com/svn/trunk@510 8350de35-f252-0410-a59d-9bf392a7b71e