New target Txt2tags (-t txt2t)
... No, seriously.
This is regular txt2tags markup, fully parsed and
then re-emitted. It does not simply run the file
through unchanged.

Why make this? Some possible benefits:

1. Testing new features. If something isn't
working right yet, run it through with this
target. If you get back different markup than
in the original file, there's an error somewhere.

2. Cleaned/normalized files. If you use a large
number of preprocs and includes, but wish to make
a cleanly marked up file for other reasons, you
can run it through this target.

Specifically, the result of -t txt2t will be a bit
different from the original file (almost all options
and macros will be expanded and removed). However,
the result of using this as a source file should
be nearly identical to using the original source
file for any other target.



git-svn-id: http://txt2tags.googlecode.com/svn/trunk@639 8350de35-f252-0410-a59d-9bf392a7b71e